refactor(astFromString): split astFromString into multiple files
