+++ src/ast/token.ts
@@ -1,0 +1,5 @@
+export enum TokenType {
+  None = 0,
+  Delimeter = 1,
+  Symbol = 2,
+}
